requirements.txt
requests==2.25.1
urllib3==1.26.3
pillow ==8.1.0